Rename GenericHostCliCommand to ServiceHostCliCommand
This reduces potential confusion as GenericHost integration is now also available in CliCommand
