+++ src/AppMotor.CliApp/AppMotor.CliApp.NuGet.props
@@ -19,7 +19,7 @@
     <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.CliVerb" />
     <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.CliCommand" />
     <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.CliParam" />
-    <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.Hosting.GenericHostCliCommand" />
+    <CommonlyUsedTypes Include="AppMotor.CliApp.CommandLine.ServiceHostCliCommand" />
 
     <CommonlyUsedTypes Include="AppMotor.CliApp.Terminals.Terminal" />
     <CommonlyUsedTypes Include="AppMotor.CliApp.Terminals.Formatting.TermText" />
